organize backend patient API into a module
