+++ microservice_patients/.gitignore
@@ -1,4 +1,4 @@
-HELP.md
+../HELP.md
 target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
