--- src/test/java/com/mariemoore/medilabo_solutions/MedilaboSolutionsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mariemoore.medilabo_solutions;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MedilaboSolutionsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
